remove BaseBuiler and CinnBuilder and merge into NetBuilder
