+++ cinn/frontend/net_builder.cc
@@ -23,6 +23,102 @@
 namespace cinn {
 namespace frontend {
 
+using common::Context;
+using common::Type;
+using hlir::framework::Operator;
+using utils::AttributeMap;
+using utils::ShapeType;
+
+NetBuilder::NetBuilder(const std::string& name) : name_(name) {}
+
+Program NetBuilder::Build(bool in_reverse) {
+  std::vector<Instruction> instrs;
+  if (in_reverse) {
+    instrs.reserve(instrs_.size());
+    for (auto it = instrs_.rbegin(); it != instrs_.rend(); it++) {
+      instrs.emplace_back(*it);
+    }
+  } else {
+    instrs = std::move(instrs_);
+  }
+
+  Program program{std::move(instrs), std::move(inputs_)};
+  program.Validate();
+  return program;
+}
+
+Placeholder NetBuilder::CreateInput(const Type& type, const std::vector<int>& shape, const std::string& id_hint) {
+  if (!id_hint.empty()) {
+    CheckVarNameValid(id_hint);
+  }
+  std::string id = id_hint.empty() ? Context::Global().NewName("placeholder") : id_hint;
+
+  inputs_.emplace_back(id);
+  auto& var  = inputs_.back();
+  var->type  = type;
+  var->shape = shape;
+  return Placeholder(var);
+}
+
+Placeholder NetBuilder::CreateInput(const Variable& var) {
+  CHECK(!var->shape.empty()) << "The input's shape is not set yet";
+  CHECK(!var->type.is_unk()) << "The input's type is not set yet";
+  inputs_.push_back(var);
+  return Placeholder(var);
+}
+
+void NetBuilder::InferShape(Instruction instr) const {
+  using ShapeFunc           = std::function<std::vector<ShapeType>(const std::vector<ShapeType>&, const AttributeMap&)>;
+  using TypeFunc            = std::function<std::vector<Type>(const std::vector<Type>&, const AttributeMap&)>;
+  const auto& op_infershape = Operator::GetAttrs<ShapeFunc>("infershape");
+  const auto& op_inferdtype = Operator::GetAttrs<TypeFunc>("inferdtype");
+
+  size_t size = instr->inputs.size();
+  std::vector<ShapeType> in_shapes(size);
+  std::vector<Type> in_types(size);
+  std::transform(
+      instr->inputs.begin(), instr->inputs.end(), in_shapes.begin(), [](const Variable& var) { return var->shape; });
+  std::transform(
+      instr->inputs.begin(), instr->inputs.end(), in_types.begin(), [](const Variable& var) { return var->type; });
+
+  auto key        = Operator::Get(instr->op_type);
+  auto out_shapes = op_infershape[key](in_shapes, instr->attrs);
+  auto out_types  = op_inferdtype[key](in_types, instr->attrs);
+
+  auto& outs            = instr->outputs;
+  size_t origin_out_num = outs.size();
+  outs.resize(out_shapes.size());
+  for (size_t i = origin_out_num; i < outs.size(); i++) {
+    outs[i] = Variable();
+  }
+  for (size_t i = 0; i < outs.size(); i++) {
+    outs[i]->shape = out_shapes[i];
+    outs[i]->type  = out_types[i];
+  }
+}
+
+Variable NetBuilder::ElementwiseOp(const std::string& op_type, const Variable& lhs, const Variable& rhs, int axis) {
+  Instruction instr(op_type, {lhs, rhs});
+  instr.SetAttr("axis", axis);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::UnaryOp(const std::string& op_type, const Variable& operand) {
+  Instruction instr(op_type, {operand});
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::BinaryOp(const std::string& op_type, const Variable& lhs, const Variable& rhs) {
+  Instruction instr(op_type, {lhs, rhs});
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
 #define NETBUILDER_UNARY_OP_DEF(func_name__, op_type__) \
   Variable NetBuilder::func_name__(const Variable& operand) { return UnaryOp(#op_type__, operand); }
 NETBUILDER_UNARY_OP_DEF(Sqrt, sqrt)
@@ -30,6 +126,35 @@ NETBUILDER_UNARY_OP_DEF(Tanh, tanh)
 NETBUILDER_UNARY_OP_DEF(Relu, relu)
 NETBUILDER_UNARY_OP_DEF(Sigmoid, sigmoid)
 NETBUILDER_UNARY_OP_DEF(Identity, identity)
+NETBUILDER_UNARY_OP_DEF(Exp, exp)
+NETBUILDER_UNARY_OP_DEF(Erf, erf)
+NETBUILDER_UNARY_OP_DEF(Rsqrt, rsqrt)
+NETBUILDER_UNARY_OP_DEF(Log, log)
+NETBUILDER_UNARY_OP_DEF(Log2, log2)
+NETBUILDER_UNARY_OP_DEF(Log10, log10)
+NETBUILDER_UNARY_OP_DEF(Floor, floor)
+NETBUILDER_UNARY_OP_DEF(Ceil, ceil)
+NETBUILDER_UNARY_OP_DEF(Round, round)
+NETBUILDER_UNARY_OP_DEF(Trunc, trunc)
+NETBUILDER_UNARY_OP_DEF(Sin, sin)
+NETBUILDER_UNARY_OP_DEF(Cos, cos)
+NETBUILDER_UNARY_OP_DEF(Tan, tan)
+NETBUILDER_UNARY_OP_DEF(Sinh, sinh)
+NETBUILDER_UNARY_OP_DEF(Cosh, cosh)
+NETBUILDER_UNARY_OP_DEF(Asin, asin)
+NETBUILDER_UNARY_OP_DEF(Acos, acos)
+NETBUILDER_UNARY_OP_DEF(Atan, atan)
+NETBUILDER_UNARY_OP_DEF(Asinh, asinh)
+NETBUILDER_UNARY_OP_DEF(Acosh, acosh)
+NETBUILDER_UNARY_OP_DEF(Atanh, atanh)
+NETBUILDER_UNARY_OP_DEF(IsNan, isnan)
+NETBUILDER_UNARY_OP_DEF(IsFinite, isfinite)
+NETBUILDER_UNARY_OP_DEF(IsInf, isinf)
+NETBUILDER_UNARY_OP_DEF(LogicalNot, logical_not)
+NETBUILDER_UNARY_OP_DEF(BitwiseNot, bitwise_not)
+NETBUILDER_UNARY_OP_DEF(Negative, negative)
+NETBUILDER_UNARY_OP_DEF(Sign, sign)
+NETBUILDER_UNARY_OP_DEF(Abs, abs)
 
 #define NETBUILDER_BINARY_OP_DEF(func_name__, op_type__) \
   Variable NetBuilder::func_name__(const Variable& lhs, const Variable& rhs) { return BinaryOp(#op_type__, lhs, rhs); }
@@ -37,6 +162,21 @@ NETBUILDER_BINARY_OP_DEF(Add, elementwise_add)
 NETBUILDER_BINARY_OP_DEF(Sub, substract)
 NETBUILDER_BINARY_OP_DEF(Div, divide)
 NETBUILDER_BINARY_OP_DEF(ReluGrad, relu_grad)
+NETBUILDER_BINARY_OP_DEF(Dot, matmul)
+NETBUILDER_BINARY_OP_DEF(FloorDiv, floor_divide)
+NETBUILDER_BINARY_OP_DEF(Mod, mod)
+NETBUILDER_BINARY_OP_DEF(FloorMod, floor_mod)
+NETBUILDER_BINARY_OP_DEF(Max, max)
+NETBUILDER_BINARY_OP_DEF(Min, min)
+NETBUILDER_BINARY_OP_DEF(Power, power)
+NETBUILDER_BINARY_OP_DEF(LogicalAnd, logical_and)
+NETBUILDER_BINARY_OP_DEF(LogicalOr, logical_or)
+NETBUILDER_BINARY_OP_DEF(LogicalXor, logical_xor)
+NETBUILDER_BINARY_OP_DEF(BitwiseAnd, bitwise_and)
+NETBUILDER_BINARY_OP_DEF(BitwiseOr, bitwise_or)
+NETBUILDER_BINARY_OP_DEF(BitwiseXor, bitwise_xor)
+NETBUILDER_BINARY_OP_DEF(LeftShift, left_shift)
+NETBUILDER_BINARY_OP_DEF(RightShift, right_shift)
 
 #define NETBUILDER_ELEMENTWISE_OP_DEF(func_name__, op_type__)                            \
   Variable NetBuilder::func_name__(const Variable& lhs, const Variable& rhs, int axis) { \
@@ -46,6 +186,253 @@ NETBUILDER_ELEMENTWISE_OP_DEF(ElementwiseAdd, elementwise_add)
 NETBUILDER_ELEMENTWISE_OP_DEF(ElementwiseMul, elementwise_mul)
 NETBUILDER_ELEMENTWISE_OP_DEF(ElementwiseDiv, divide)
 NETBUILDER_ELEMENTWISE_OP_DEF(ElementwiseSub, substract)
+
+const std::vector<Variable>& NetBuilder::CustomInstr(const std::string& type,
+                                                     const std::vector<Variable>& inputs,
+                                                     const AttributeMap& attrs) {
+  Instruction instr(type);
+  instr.SetInputs(inputs);
+  for (auto& kv : attrs) {
+    instr.SetAttr(kv.first, kv.second);
+  }
+
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutputs();
+}
+
+Variable NetBuilder::Compare(const Variable& lhs, const Variable& rhs, ComparisonKind kind) {
+  switch (kind) {
+    case ComparisonKind::kEq:
+      return BinaryOp("equal", lhs, rhs);
+    case ComparisonKind::kNe:
+      return BinaryOp("not_equal", lhs, rhs);
+    case ComparisonKind::kGe:
+      return BinaryOp("greater_equal", lhs, rhs);
+    case ComparisonKind::kGt:
+      return BinaryOp("greater", lhs, rhs);
+    case ComparisonKind::kLe:
+      return BinaryOp("less_equal", lhs, rhs);
+    case ComparisonKind::kLt:
+      return BinaryOp("less", lhs, rhs);
+    default:
+      LOG(FATAL) << "unknown comparison kind";
+  }
+}
+
+std::vector<Variable> NetBuilder::Split(const Variable& operand, const std::vector<int>& num_or_sections, int axis) {
+  Instruction instr("split", {operand});
+  instr.SetAttr("num_or_sections", num_or_sections);
+  instr.SetAttr("axis", axis);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutputs();
+}
+
+Variable NetBuilder::Concat(const std::vector<Variable>& input_vars, int axis) {
+  Instruction instr("concat", input_vars);
+  instr.SetAttr("axis", axis);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::Reduce(const Variable& operand, ReduceKind kind, const std::vector<int>& dim, bool keep_dim) {
+  auto reduce_func = [&](const std::string& op_type) {
+    Instruction instr(op_type, {operand});
+    instr.SetAttr("dim", dim);
+    instr.SetAttr("keep_dim", keep_dim);
+    InferShape(instr);
+    AppendInstruction(instr);
+    return instr.GetOutput(0);
+  };
+
+  switch (kind) {
+    case ReduceKind::kSum:
+      return reduce_func("reduce_sum");
+    case ReduceKind::kProd:
+      return reduce_func("reduce_prod");
+    case ReduceKind::kMax:
+      return reduce_func("reduce_max");
+    case ReduceKind::kMin:
+      return reduce_func("reduce_min");
+    case ReduceKind::kAll:
+      return reduce_func("reduce_all");
+    case ReduceKind::kAny:
+      return reduce_func("reduce_any");
+    default:
+      LOG(FATAL) << "unknown reduction kind";
+  }
+}
+
+Variable NetBuilder::FillConstant(
+    const std::vector<int>& shape, float value, const std::string& name, const std::string& dtype, bool force_cpu) {
+  Instruction instr("fill_constant");
+  instr.SetInputs({});
+  instr.SetAttr("shape", shape);
+  instr.SetAttr("value", value);
+  instr.SetAttr("dtype", dtype);
+  instr.SetAttr("force_cpu", force_cpu);
+
+  InferShape(instr);
+  AppendInstruction(instr);
+  auto out = instr.GetOutput(0);
+  out.set_id(name);
+  return out;
+}
+
+Variable NetBuilder::BroadcastTo(const Variable& operand, const std::vector<int>& out_shape) {
+  auto x_shape_size = operand->shape.size();
+  auto y_shape_size = out_shape.size();
+  CHECK_GT(x_shape_size, 0) << "Cannot broadcast a empty operand " << operand->id << " to "
+                            << cinn::utils::Join(out_shape, ",");
+  CHECK_LE(x_shape_size, y_shape_size) << "The broadcast_p's input shape dimension should less than the output's, "
+                                       << "but here (" << x_shape_size << " > " << y_shape_size << ").";
+
+  VLOG(4) << "Try broadcast " << operand->id << " from shape (" << cinn::utils::Join(operand->shape, ",")
+          << ") to shape (" << cinn::utils::Join(out_shape, ",") << ")";
+
+  std::vector<int> broadcast_axes(x_shape_size, 0);
+  if (x_shape_size > 1) {
+    for (int i = 1; i <= x_shape_size; ++i) {
+      CHECK((out_shape[y_shape_size - i] == operand->shape[x_shape_size - i]) ||
+            (operand->shape[x_shape_size - i] == 1))
+          << "We cannot broadcast from shape (" << cinn::utils::Join(operand->shape, ",") << ") to shape ("
+          << cinn::utils::Join(out_shape, ",") << ")";
+      broadcast_axes[x_shape_size - i] = y_shape_size - i;
+    }
+  } else {
+    int axis     = -1;
+    auto x_shape = operand->shape.at(0);
+    if (x_shape == 1) {
+      // Can broadcast directly, default axis 0
+      axis = 0;
+    } else {
+      // The broadcast axes is the index of the shape in out_shape when the input dimension is 1
+      for (int i = 0; i < y_shape_size; ++i) {
+        if (out_shape[i] == x_shape) {
+          axis = i;
+          break;
+        }
+      }
+      CHECK_NE(axis, -1) << "When we broadcast a 1-dimension shape, the number should contained in the out_shape. "
+                         << "We cannot broadcast from shape (" << cinn::utils::Join(operand->shape, ",")
+                         << ") to shape (" << cinn::utils::Join(out_shape, ",") << ")";
+    }
+    broadcast_axes[0] = axis;
+  }
+
+  return BroadcastTo(operand, out_shape, broadcast_axes);
+}
+
+Variable NetBuilder::BroadcastTo(const Variable& operand,
+                                 const std::vector<int>& out_shape,
+                                 const std::vector<int>& broadcast_axes) {
+  Instruction instr("broadcast_to", {operand});
+  instr.SetAttr("out_shape", out_shape);
+  instr.SetAttr("broadcast_axes", broadcast_axes);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::Reshape(const Variable& operand, const std::vector<int>& shape) {
+  Instruction instr("reshape", {operand});
+  instr.SetAttr("shape", shape);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::Transpose(const Variable& operand, const std::vector<int>& axis) {
+  Instruction instr("transpose", {operand});
+  instr.SetAttr("axis", axis);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::Slice(const Variable& operand,
+                           const std::vector<int>& axes,
+                           const std::vector<int>& starts,
+                           const std::vector<int>& ends,
+                           const std::vector<int>& infer_flags,
+                           const std::vector<int>& strides) {
+  Instruction instr("slice", {operand});
+  instr.SetAttr("axes", axes);
+  instr.SetAttr("starts", starts);
+  instr.SetAttr("ends", ends);
+  instr.SetAttr("infer_flags", infer_flags);
+  instr.SetAttr("strides", strides);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::SliceAssign(const Variable& input,
+                                 const Variable& assign,
+                                 const std::vector<int>& axes,
+                                 const std::vector<int>& starts,
+                                 const std::vector<int>& ends,
+                                 const std::vector<int>& strides) {
+  Instruction instr("slice_assign", {input, assign});
+  instr.SetAttr("axes", axes);
+  instr.SetAttr("starts", starts);
+  instr.SetAttr("ends", ends);
+  instr.SetAttr("strides", strides);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::Reverse(const Variable& operand, const std::vector<int>& axis) {
+  Instruction instr("reverse", {operand});
+  instr.SetAttr("axis", axis);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::Select(const Variable& condition, const Variable& true_value, const Variable& false_value) {
+  Instruction instr("select", {condition, true_value, false_value});
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::IndexSelect(const Variable& operand, const Variable& index, int axis) {
+  Instruction instr("index_select", {operand, index});
+  instr.SetAttr("axis", axis);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::ScatterAssign(const Variable& operand, const Variable& updates, const Variable& index, int axis) {
+  Instruction instr("scatter_assign", {operand, updates, index});
+  instr.SetAttr("axis", axis);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::ScatterAdd(const Variable& operand, const Variable& updates, const Variable& index, int axis) {
+  Instruction instr("scatter_add", {operand, updates, index});
+  instr.SetAttr("axis", axis);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::IsClose(const Variable& x, const Variable& y, float rtol, float atol, bool equal_nan) {
+  Instruction instr("isclose", {x, y});
+  instr.SetAttr("rtol", rtol);
+  instr.SetAttr("atol", atol);
+  instr.SetAttr("equal_nan", equal_nan);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
 
 Variable NetBuilder::Mul(const Variable& a, const Variable& b, int x_num_col_dims, int y_num_col_dims) {
   Instruction instr("mul", {a, b});
@@ -98,6 +485,32 @@ Variable NetBuilder::Cast(const Variable& operand, const std::string& dtype) {
 Variable NetBuilder::Squeeze(const Variable& operand, const std::vector<int>& axes) {
   Instruction instr("squeeze", {operand});
   instr.SetAttr("axes", axes);
+  InferShape(instr);
+  AppendInstruction(instr);
+  return instr.GetOutput(0);
+}
+
+Variable NetBuilder::Conv(const Variable& lhs,
+                          const Variable& rhs,
+                          const std::vector<int>& strides,
+                          const std::vector<int>& paddings,
+                          const std::vector<int>& dilations,
+                          int groups,
+                          const std::string& conv_type,
+                          const std::string& data_format,
+                          const std::string& padding_algorithm,
+                          const std::vector<int>& output_shape) {
+  Instruction instr("conv2d");
+  instr.SetInputs({lhs, rhs});
+  instr.SetAttr("stride", strides);
+  instr.SetAttr("padding", paddings);
+  instr.SetAttr("dilation", dilations);
+  instr.SetAttr("groups", groups);
+  instr.SetAttr("conv_type", conv_type);
+  instr.SetAttr("data_format", data_format);
+  instr.SetAttr("padding_algorithm", padding_algorithm);
+  instr.SetAttr("output_shape", output_shape);
+
   InferShape(instr);
   AppendInstruction(instr);
   return instr.GetOutput(0);
@@ -503,14 +916,6 @@ Variable NetBuilder::Matmul(const Variable& x, const Variable& y, bool trans_x, 
   }
 
   return out;
-}
-
-Variable NetBuilder::ElementwiseOp(const std::string& op_type, const Variable& lhs, const Variable& rhs, int axis) {
-  Instruction instr(op_type, {lhs, rhs});
-  instr.SetAttr("axis", axis);
-  InferShape(instr);
-  AppendInstruction(instr);
-  return instr.GetOutput(0);
 }
 
 }  // namespace frontend
