+++ cinn/frontend/pass/transpose_collapsing.cc
@@ -99,7 +99,7 @@ class TransposeCollapsingPass : public ProgramPass {
     }
     FoldingTransposeHorizontal(all_transpose, fetch_ids, in2instr, out2instr, &remove_instrs);
 
-    CinnBuilder builder("transpose_collapsing_builder");
+    NetBuilder builder("transpose_collapsing_builder");
     for (auto& var : program->GetInputs()) {
       builder.CreateInput(var);
     }
